Class MabedeError is in its own module
