+++ mabede/src/mabede-error.js
@@ -1,0 +1,10 @@
+class MabedeError {
+	constructor(statusCode, content) {
+		this.statusCode = statusCode;
+		this.content = content;
+	}
+
+	toString = () => `[${this.statusCode}] ${this.content}`;
+}
+
+module.exports = MabedeError;
